Israa

+++ app/Http/Controllers/UserController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 use Symfony\Component\HttpFoundation\Response;
 use Illuminate\Support\Facades\Hash;
+use Spatie\Permission\Models\Role;
 use Str;
 
 class UserController extends Controller
@@ -16,7 +18,7 @@ class UserController extends Controller
 
     public function __construct()
     {
-     $this-> authorizeResource(User::class, 'user');
+         $this-> authorizeResource(User::class, 'user');
     }
 
 
@@ -25,11 +27,12 @@ class UserController extends Controller
         $users = User::all();
         return response()->view('cms.users.index', ['users' => $users]);
     }
-
-
     public function create()
     {
-        return response()->view('cms.users.create');
+        $roles=Role::where('guard_name','user')->get();
+
+        $countries =  DB::table('countries')->get();
+        return response()->view('cms.users.create', ['countries' => $countries,'roles'=>$roles]);
     }
 
 
@@ -41,10 +44,18 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        // dd($request->input('user_role'));
         $validator = Validator($request->all(), [
             'name' => 'required|string|min:3',
             'email' => 'required|email|unique:users,email',
             'user_type' => 'required',
+
+            'country_id' => 'required',
+            'city_id' => 'required',
+            'region' => 'required|string|min:3',
+         'user_role'=>'required|numeric|exists:roles,id',
+
+
             'image' => 'required|image|mimes:png,jpg,jpeg',
         ]);
         if (!$validator->fails()) {
@@ -52,29 +63,50 @@ class UserController extends Controller
             $user->name = $request->input('name');
             $user->email = $request->input('email');
             $user->user_type = $request->input('user_type');
-            // $user->password = Hash::make(12345);
-            $user->password = Hash::make(Str::random(8));
+            $user->country_id = $request->input('country_id');
+            $user->city_id = $request->input('city_id');
+            $user->region = $request->input('region');
 
+            $user->password = Hash::make(Str::random(8));
+            $user->admin_id=auth()->user()->getAuthIdentifier();
             if ($request->hasFile('image')) {
-                $file = $request->file('image');
-                $imagetitle =  time() . '_user_image.' . $file->getClientOriginalExtension();
-                $status = $request->file('image')->storePubliclyAs('images/users', $imagetitle);
-                $imagePath = 'images/users/' . $imagetitle;
-                $user->image = $imagePath;
+                $imagetitle =  time() . '_'. str_replace(' ','',$user->name).'.'. $request->file('image')->extension();
+                $request->file('image')->storePubliclyAs('users', $imagetitle,['disk'=>'public']);
+                $user->image = 'users/'.$imagetitle;
             }
             $isSaved = $user->save();
+             if($isSaved){
+               $role= Role::where('id',$request->get('user_role'))->first();
+                $user->assignRole($role);
+            }
+
+
+
             return response()->json(
-             ['message' => $isSaved ? 'Saved successfully' : 'Save failed!'],
-              $isSaved ? Response::HTTP_CREATED : Response::HTTP_BAD_REQUEST);
+                ['message' => $isSaved ? 'Saved successfully' : 'Save failed!'],
+                $isSaved ? Response::HTTP_CREATED : Response::HTTP_BAD_REQUEST
+            );
         } else {
             return response()->json(
                 ['message' => $validator->getMessageBag()->first()],
-                Response::HTTP_BAD_REQUEST,);
+                Response::HTTP_BAD_REQUEST,
+            );
         }
     }
     public function edit(User $user)
+
+
+
+
+
+
     {
-        return response()->view('cms.users.edit', ['user' => $user]);
+
+        $roles=Role::where('guard_name','user')->get();
+
+        $countries =  DB::table('countries')->get();
+        $cities = DB::table('cities')->get();
+        return response()->view('cms.users.edit', ['user' => $user,'countries'=>$countries,'cities'=>$cities,'roles'=>$roles]);
     }
 
     /**
@@ -84,32 +116,49 @@ class UserController extends Controller
      * @param  \App\Models\Admin  $admin
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, User $user)
+    public function update(Request $request, $id)
     {
         $validator = Validator($request->all(), [
-            'name' => 'nullable|string|min:3',
-            'email' => 'nullable|email|unique:users,email,' . $user->id,
-            'user_type' => 'required',
-            'image' => 'nullable', 'image|mimes:png,jpg,jpeg',
+            // 'name' => 'nullable|string|min:3',
+            // 'email' => 'nullable|email|unique:users,email,' ,
+            // 'user_type' => 'required',
+            // 'country_id' => 'required',
+            // 'city_id' => 'required',
+            // 'region' => 'required|string|min:3',
+            // 'image' => 'nullable', 'image|mimes:png,jpg,jpeg',
+
         ]);
         if (!$validator->fails()) {
+            $user = User::findOrFail($id);
             $user->name = $request->input('name');
             $user->email = $request->input('email');
             $user->user_type = $request->input('user_type');
+
+            $user->country_id = $request->input('country_id');
+            $user->city_id = $request->input('city_id');
+            $user->region = $request->input('region');
+            $user->admin_id=auth()->user()->getAuthIdentifier();
+
+
+
             if ($request->hasFile('image')) {
-                //Delete admin previous image.
                 Storage::delete($user->image);
-                //Save new image.
-                $file = $request->file('image');
-                $imageName = time() . '_user_image.' . $file->getClientOriginalExtension();
-                $request->file('image')->storePubliclyAs('images/users', $imageName);
-                $imagePath = 'images/users/' . $imageName;
-                $user->image = $imagePath;
+                $imagetitle =  time() . '_'. str_replace(' ','',$user->name).'.'. $request->file('image')->extension();
+                $request->file('image')->storePubliclyAs('users', $imagetitle,['disk'=>'public']);
+                $user->image = 'users/'.$imagetitle;
             }
+
             $isSaved = $user->save();
+            if($isSaved){
+              $role= Role::where('id',$request->get('user_role'))->first();
+              $saved= $user->syncRoles($role);
+            // $user->assignRole(Role::findById($request->get('role_id')))->first();
+
+           }
             return response()->json(
                 ['message' => $isSaved ? 'Updated Successfully' : 'Update failed!'],
-                $isSaved ? Response::HTTP_OK : Response::HTTP_BAD_REQUEST);
+                $isSaved ? Response::HTTP_OK : Response::HTTP_BAD_REQUEST
+            );
         } else {
             return response()->json(['message' => $validator->getMessageBag()->first()], Response::HTTP_BAD_REQUEST);
         }
@@ -128,7 +177,7 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         $deleted = $user->delete();
-        if ($deleted) {
+        if ($deleted && $user->image!=null) {
             Storage::delete($user->image);
         }
         return response()->json(
